Added all the necessary files for the program to work
The software allows users to enter mathematical functions and find their graphs on the plane. The application offers zooming capabilities, easy graph navigation, and the identification of key points such as extrema and intersections.
The amount of space that the software takes up on the hard disk is approximately 10 MB.
The Android operating system and Java programming language were used to write the program. The application is easily integrated with Android devices, making it accessible to a wide range of users.
The input data for the program is a mathematical function entered by the user, as well as scaling and navigation parameters.
The result of the application is a graph of the entered function on the plane, which the user can explore and analyze.
The application provides a user-friendly interface for entering mathematical functions and displaying their graphs, making it a useful tool for students, teachers, and anyone interested in mathematics and function analysis.